winner logic is done
